Time: 18 ms (60.82%), Space: 45.1 MB (14.87%) - LeetHub

+++ 0018-4sum/0018-4sum.java
@@ -15,7 +15,7 @@ class Solution {
             int k = n-1;
             while( j<k)
             {
-                long sum =  ((long)nums[i] + (long)nums[j] +(long) nums[l] +  (long)nums[k]);
+                long sum = (long)nums[i] + nums[j] + nums[l] +  nums[k];
                 long target = (long) tar;
                 if( sum < target){
                     j++;
